Fix compiler warnings. Rename DisplayCompletionOptions() - this can be used generally when showing informational messages. Protect for non-existing Colorizer. Put TextInfo object into Context. Store the most recently entered key such that ROOT can access it. More lower-case locals. Implement tab completion handling, resize "commands" (relevant on Windows). Don't delete Colorizer, we don't own it. Check end == -1 in SetColor().
git-svn-id: https://root.cern.ch/svn/root/branches/dev/axel@39295 27541ba8-7e3a-0410-8455-c3a389f83636

+++ textinput/History.cpp
@@ -73,7 +73,7 @@ namespace textinput {
       // Prune!
       fHistFile.seekp(0);
       size_t nPrune = fPruneLength;
-      if (nPrune == kPruneLengthMaxDepth) {
+      if (nPrune == (size_t)kPruneLengthMaxDepth) {
         nPrune = fMaxDepth;
       }
       for (size_t Idx = fEntries.size() - nPrune,
